requirements.txt
passlib[bcrypt]
python-jose[cryptography]
apscheduler
pandas
matplotlib
psycopg2
sqlalchemy
fastapi[all]